Moved code down into models
